requirements.txt
django
gunicorn
requests
django-heroku
cmake
Pillow
face_recognition
opencv-python-headless